requirements.txt
faiss_cpu==1.11.0
numpy>=1.26.0
openai==1.78.1
python-dotenv==1.1.0
streamlit==1.45.1
